% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{data}
\name{counties_us}
\alias{counties_us}
\title{US counties as GeoJSON}
\format{vector of GeoJSON strings, one for each county in the USA}
\description{
US counties as GeoJSON
}
\keyword{datasets}
