% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{data}
\name{counties_us_df}
\alias{counties_us_df}
\title{US counties as GeoJSON, in a data.frame with metadata}
\format{data.frame of GeoJSON strings, one for each county in the USA}
\description{
US counties as GeoJSON, in a data.frame with metadata
}
\keyword{datasets}
