% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{data}
\name{gsp_congressional_df}
\alias{gsp_congressional_df}
\title{US congressional districts as GeoJSON, in a data.frame with metadata}
\format{data.frame of GeoJSON strings, one for each congressional district.
Rows: 437. Columns: 8. Data here is 20m resolution}
\description{
US congressional districts as GeoJSON, in a data.frame with metadata
}
\details{
Columns:
\itemize{
\item geo_id
\item state: name of the state
\item state_code: state FIPS code
\item cd
\item name
\item lsad
\item censusarea
\item geojson: GeoJSON as character
}
}
\keyword{datasets}
