% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{data}
\name{gsp_counties}
\alias{gsp_counties}
\title{US counties as GeoJSON}
\format{named list of GeoJSON strings, one for each county in the USA.
top level list names are states, then counties within each state.
Length: 52. Total number of counties: 3221}
\description{
US counties as GeoJSON
}
\keyword{datasets}
