% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{data}
\name{gsp_states}
\alias{gsp_states}
\title{US states as GeoJSON}
\format{named list of GeoJSON strings, one for each state in the USA.
top level list names are states. Length: 52. Data here is 20m resolution}
\description{
US states as GeoJSON
}
\keyword{datasets}
