% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{data}
\name{gsp_us}
\alias{gsp_us}
\title{US outline as GeoJSON}
\format{named list of GeoJSON strings, one for each of three levels
of resolution:
\itemize{
\item us_20m: least detailed, resolution of 1:20,000,000
\item us_5m: medium detail, resolution of 1:5,000,000
\item us_500k: most detailed, resolution of 1:500,000
}}
\description{
US outline as GeoJSON
}
\keyword{datasets}
