% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{package}
\name{geojsonpolys-package}
\alias{geojsonpolys-package}
\alias{geojson}
\title{geojsonpolys}
\description{
GeoJSON Polygons. Data provided in this package is of
the smallest size of the options provided, and the least resolution.
If you want
}
\examples{
length(gsp_counties)
gsp_counties[1]
gsp_counties$alabama
gsp_counties$alabama$walker
head(gsp_counties_df)
}
\author{
Scott Chamberlain
}
\keyword{package}
