% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonpolys-package.R
\docType{package}
\name{geojsonpolys-package}
\alias{geojsonpolys-package}
\alias{geojson}
\title{geojsonpolys}
\description{
GeoJSON Polygons
}
\examples{
length(counties_us)
counties_us[1]
counties_us_df
}
\author{
Scott Chamberlain
}
\keyword{package}
